backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
requests
pydantic
python-dotenv
pytest
pytest-asyncio
